pyproject.toml
[project]
name = "predicting-temperature"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "marimo>=0.16.0",
    "mlflow[databricks]>=3.0.1",
    "python-dotenv>=1.1.1",
    "seaborn>=0.13.2",
]
